requirements.txt
tkinter
Pillow
winshell
auto-py-to-exe